Bug: Refactoring the task object and how it is written to files
I realized the format in the todo.txt file is wrong for completed tasks. The
code has been updated to support both my old broken layout and the correct
one. It also should rewrite all the old bad values the next time you read in
all the tasks and have it write out again since it does a full replace on the
files.